requirements.txt
cachetools==5.3.3
deprecation==2.1.0
grpcio==1.62.1
packaging==24.0
protobuf==4.25.3
pyee==11.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b97
typing_extensions==4.10.0

pandas==2.2.3
ta~=0.11.0
PyYAML~=6.0.1
requests~=2.31.0
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-Sockets==0.2.1
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4

psutil==7.0.0
pycparser==2.22
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
tzdata==2025.2
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2